fix [forkedTask]: move func isForkingTask/isRunAsMeTask from scheduler-api to scheduler-server, because it's co-decided by global configuration which is only available on server, scheduler-api may exec on client.
