+++ scheduler/scheduler-api/src/test/java/org/ow2/proactive/scheduler/common/job/factories/TestJobFactory.java
@@ -127,8 +127,8 @@ public class TestJobFactory {
         assertEquals("1", ((JavaTask) task1).getArgument("sleepTime"));
         assertEquals("1", ((JavaTask) task1).getArgument("number"));
         assertEquals("org.ow2.proactive.scheduler.examples.WaitAndPrint", ((JavaTask) task1).getExecutableClassName());
-        // Now whether the task is executed in a forked JVM, it's decided by the parameter fork and runasme, instead of forkEnvironment or walltime, since 3.12.
-        assertEquals(Task.DEFAULT_TASK_FORK, task1.isForkingTask());
+        // Now whether the task is executed in a forked JVM, it's specified by the parameter fork and runasme, instead of forkEnvironment or walltime, since 3.12.
+        assertEquals(null, task1.isFork());
         assertEquals(null, task1.getForkEnvironment());
 
         //Check task 2 properties
@@ -168,8 +168,8 @@ public class TestJobFactory {
         assertEquals("21", ((JavaTask) task2).getArgument("number"));
         assertEquals("/bin/java/jdk1.5", ((JavaTask) task2).getArgument("test"));
         assertEquals("org.ow2.proactive.scheduler.examples.WaitAndPrint", ((JavaTask) task2).getExecutableClassName());
-        // Now whether the task is executed in a forked JVM, it's decided by the parameter fork and runasme, instead of forkEnvironment or walltime, since 3.12.
-        assertEquals(Task.DEFAULT_TASK_FORK, task2.isForkingTask());
+        // Now whether the task is executed in a forked JVM, it's specified by the parameter fork and runasme, instead of forkEnvironment or walltime, since 3.12.
+        assertEquals(null, task2.isFork());
         assertEquals(false, task2.isWallTimeSet());
         assertEquals("/bin/java/jdk1.5", task2.getForkEnvironment().getJavaHome());
         assertEquals("/bin/java/jdk1.5/toto", task2.getForkEnvironment().getWorkingDir());
